requirements.txt
requests
Flask
Flask[async]
python-dotenv
gunicorn
azure-data-tables
azure-storage-blob
langchain
langchain-openai
langchain-deepseek
langchain-community
langgraph<0.4.0,>=0.2.56
langsmith<0.4.0,>=0.1.63
xmltodict
pydantic
duckduckgo-search
playwright
lxml
beautifulsoup4